Create 16x16 squares with red border
First, created a square having width and height
of 10px with a red border then used a for loop
to create 16 more squares and 16 rows creating a
16x16 grid.

Created the grid by leveraging for loop. Inside
loop i created each row divs and another for loop
for square divs. Used flexbox to align them
in a grid pattern.