Add a hover effect to the squares
Added an effect to the squares where if you mouse
over them they change to a different color and
persist like that making it possible to draw a
sketch.

Achieved this by leveraging event listener. First,
i created a variable boxes that has a node list of
all the squares. Then used for each on that
variable and implemented event listeners on each
box.