Add: logic to button; Refactor: create, del grid
Added logic to button, it nows shows a prompt to
get user input for number of squares per side.

Also refactored code that creates grid into a
function called createGrid.

Created a function called deleteGrid, this is to
ensure that whenever user inputs a value, the
current grid gets deleted and a new grid is
created. Otherwise they would just add to the
current grid.