Add reset button and is cleaner and readable
Refactored some of the code to be inside a
function. Now Most of the code is in functions.
startGrid is the beginning of the program and
hover is for the hovering effect of squares.
Rather than creating grid again when user chooses
an input, startGrid is called again with the user
choice as argument. deleteGrid is the same.

Added a RESET button. It reset the current grid
with the same size.